repl: move `complete` to its own file
