+++ lib/internal/repl/utils.js
@@ -46,6 +46,14 @@ const {
   getStringWidth,
   inspect,
 } = require('internal/util/inspect');
+const vm = require('vm');
+const path = require('path');
+const CJSModule = require('internal/modules/cjs/loader').Module;
+
+const builtinLibs = ArrayPrototypeFilter(
+  CJSModule.builtinModules,
+  (e) => !StringPrototypeStartsWith(e, '_'),
+);
 
 let debug = require('internal/util/debuglog').debuglog('repl', (fn) => {
   debug = fn;
@@ -742,6 +750,27 @@ function setupReverseSearch(repl) {
   return { reverseSearch };
 }
 
+let nextREPLResourceNumber = 1;
+// This prevents v8 code cache from getting confused and using a different
+// cache from a resource of the same name
+function getREPLResourceName() {
+  return `REPL${nextREPLResourceNumber++}`;
+}
+
+const globalBuiltins = new SafeSet(vm.runInNewContext('Object.getOwnPropertyNames(globalThis)'));
+
+let replFilename;
+try {
+  // Hack for require.resolve("./relative") to work properly.
+  replFilename = path.resolve('repl');
+} catch {
+  // path.resolve('repl') fails when the current working directory has been
+  // deleted.  Fall back to the directory name of the (absolute) executable
+  // path.  It's not really correct but what are the alternatives?
+  const dirname = path.dirname(process.execPath);
+  replFilename = path.resolve(dirname, 'repl');
+}
+
 module.exports = {
   REPL_MODE_SLOPPY: Symbol('repl-sloppy'),
   REPL_MODE_STRICT,
@@ -749,4 +778,9 @@ module.exports = {
   kStandaloneREPL: Symbol('kStandaloneREPL'),
   setupPreview,
   setupReverseSearch,
+  getREPLResourceName,
+  kContextId: Symbol('kContextId'),
+  globalBuiltins,
+  builtinLibs,
+  replFilename,
 };
